+++ shared/editor/diagram/index.ts
@@ -1,4 +1,5 @@
 export * from './DiagramRenderCoordinator';
+export * from './DiagramRenderIntentStore';
 export * from './editorRenderer';
 export * from './externalDiagramNodes';
 export * from './languages';